Update: Backend, reasoning, LNN core and frontend enhancements
- Improved backend service and capability switching
- Added reasoning_internal module, removed deprecated reasoning_stubs
- Enhanced LNN core, selflnn, and quaternion modules
- Updated frontend CSS, index.html, api-service, data-engine, training-center
- Updated CMakeLists, README, platform utils and hardware detector
- Added .gitignore rules for backups/ and selflnn_config.json

+++ include/selflnn/core/quaternion_enhanced.h
@@ -157,7 +157,7 @@ static inline void quat_to_euler(const Quaternion q, float* roll, float* pitch, 
 
     float sinp = 2.0f * (q.w * q.y - q.z * q.x);
     if (fabsf(sinp) >= 1.0f)
-        *pitch = copysignf(M_PI / 2.0f, sinp);
+        *pitch = copysignf((float)M_PI / 2.0f, sinp);
     else
         *pitch = asinf(sinp);
